event-consumer/requirements.txt
kafka-python==2.0.2
mysql-connector-python==9.0.0
pytest==8.3.3
